// src/app/modules/auth/auth.interface.ts
import { ENUM_USER_ROLE } from '../../../enums/user';

export type ILoginUser = {
  phoneNumber: number;
  password: string;
};

export type ILoginUserResponse = {
  accessToken: string;
  refreshToken?: string;
  needsPasswordChange: boolean;
};

export type IRefreshTokenResponse = {
  accessToken: string;
};

export type IVerifiedLoginUser = {
  phoneNumber: number;
  role: ENUM_USER_ROLE;
};

export type IChangePassword = {
  oldPassword: string;
  newPassword: string;
};